+++ src/Aggregations/Types/MinBucket.php
@@ -1,0 +1,33 @@
+<?php
+
+namespace Codeart\OpensearchLaravel\Aggregations\Types;
+
+use Codeart\OpensearchLaravel\Interfaces\OpenSearchQuery;
+
+class MinBucket implements OpenSearchQuery, AggregationType
+{
+    public function __construct(
+        private readonly string $bucketsPath,
+        private readonly ?string $gapPolicy
+    ){}
+
+    public static function make(string $bucketsPath, ?string $gapPolicy = null): self
+    {
+        return new self($bucketsPath, $gapPolicy);
+    }
+
+    public function toOpenSearchQuery(): array
+    {
+        $query = [
+            'min_bucket' => [
+                'buckets_path' => $this->bucketsPath,
+            ]
+        ];
+
+        if (!is_null($this->gapPolicy)) {
+            $query['min_bucket']['gap_policy'] = $this->gapPolicy;
+        }
+
+        return $query;
+    }
+}